requirements.txt
Flask==2.0.1
flask-login==0.6.2
flask-mongoengine==1.0.0
python-dotenv==1.0.0
pymongo==4.6.1
mongoengine==0.27.0
Werkzeug==2.0.1
markdown2==2.4.12
gunicorn==21.2.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-WTF==1.2.2
python-jose==3.3.0
email_validator==2.2.0
Jinja2==3.1.4
itsdangerous==2.2.0
MarkupSafe==2.1.5
